fix: bugs in notifications based on clients notice

+++ src/tweets/tweets.repository.ts
@@ -341,12 +341,31 @@ export class TweetsRepository {
     await prismaClient.retweet.deleteMany({
       where: { tweetId },
     });
-    await prismaClient.notification.deleteMany({
-      where: { tweetId },
-    });
     await prismaClient.like.deleteMany({
       where: { tweetId },
     });
+
+    const result = await prismaClient.notification.findMany({
+      where: { tweetId },
+      select: { receiverId: true },
+    });
+
+    if (result.length > 0) {
+      await prismaClient.notification.deleteMany({
+        where: { tweetId },
+      });
+      const counts = await Promise.all(
+        result.map((user) =>
+          prismaClient.notification.count({
+            where: { receiverId: user.receiverId, seen: false },
+          }),
+        ),
+      );
+      return result.map((row, index) => ({
+        receiverId: row.receiverId,
+        unseenCount: counts[index],
+      }));
+    }
   }
 
   mapToDetailedTweetDto(tweet: DetailedTweetWithIncludes): TweetDto & { replyToTweet?: TweetDto } {
